完成了job下execute_job的类的继承与方法的复用
