+++ pipeline_plugins/components/collections/sites/open/job/execute_task/execute_task_base.py
@@ -12,7 +12,9 @@ specific language governing permissions and limitations under the License.
 """
 
 from functools import partial
+from copy import deepcopy
 
+from django.utils import translation
 from django.utils.translation import ugettext_lazy as _
 
 from pipeline.core.flow.io import (
@@ -23,6 +25,12 @@ from pipeline.core.flow.io import (
     BooleanItemSchema,
 )
 from pipeline_plugins.components.collections.sites.open.job import JobService
+from pipeline_plugins.components.utils import (
+    get_job_instance_url,
+    get_node_callback_url,
+    loose_strip,
+    get_biz_ip_from_frontend,
+)
 from gcloud.conf import settings
 from gcloud.utils.handlers import handle_api_error
 
@@ -34,6 +42,11 @@ job_handle_api_error = partial(handle_api_error, __group_name__)
 
 
 class JobExecuteTaskServiceBase(JobService):
+    """
+    JobExecuteTaskServiceBase类是job.execute_task所有legacy与v1.0版本的父类;
+    由于两个版本仅再前端处理逻辑上不同，所以两个版本的后端代码可以直接复用JobExecuteTaskServiceBase类
+    """
+
     need_get_sops_var = True
 
     def inputs_format(self):
@@ -91,6 +104,90 @@ class JobExecuteTaskServiceBase(JobService):
                 ),
             ),
         ]
+
+    def execute(self, data, parent_data):
+        executor = parent_data.get_one_of_inputs("executor")
+        client = get_client_by_user(executor)
+        client.set_bk_api_ver("v2")
+        if parent_data.get_one_of_inputs("language"):
+            setattr(client, "language", parent_data.get_one_of_inputs("language"))
+            translation.activate(parent_data.get_one_of_inputs("language"))
+
+        biz_cc_id = data.get_one_of_inputs("biz_cc_id", parent_data.inputs.biz_cc_id)
+        original_global_var = deepcopy(data.get_one_of_inputs("job_global_var"))
+        global_vars = []
+        ip_is_exist = data.get_one_of_inputs("ip_is_exist")
+        biz_across = data.get_one_of_inputs("biz_across")
+
+        for _value in original_global_var:
+            val = loose_strip(_value["value"])
+            # category为3,表示变量类型为IP
+            if _value["category"] == 3:
+                if biz_across:
+                    result, ip_list = get_biz_ip_from_frontend(
+                        ip_str=val,
+                        executor=executor,
+                        biz_cc_id=biz_cc_id,
+                        data=data,
+                        logger_handle=self.logger,
+                        is_across=True,
+                        ip_is_exist=ip_is_exist,
+                        ignore_ex_data=True,
+                    )
+
+                    # 匹配不到云区域IP格式IP，尝试从当前业务下获取
+                    if not result:
+                        result, ip_list = get_biz_ip_from_frontend(
+                            ip_str=val,
+                            executor=executor,
+                            biz_cc_id=biz_cc_id,
+                            data=data,
+                            logger_handle=self.logger,
+                            is_across=False,
+                            ip_is_exist=ip_is_exist,
+                        )
+
+                    if not result:
+                        return False
+                else:
+                    result, ip_list = get_biz_ip_from_frontend(
+                        ip_str=val,
+                        executor=executor,
+                        biz_cc_id=biz_cc_id,
+                        data=data,
+                        logger_handle=self.logger,
+                        is_across=False,
+                        ip_is_exist=ip_is_exist,
+                    )
+                    if not result:
+                        return False
+
+                if ip_list:
+                    global_vars.append({"name": _value["name"], "ip_list": ip_list})
+            else:
+                global_vars.append({"name": _value["name"], "value": val})
+
+        job_kwargs = {
+            "bk_biz_id": biz_cc_id,
+            "bk_job_id": data.get_one_of_inputs("job_task_id"),
+            "global_vars": global_vars,
+            "bk_callback_url": get_node_callback_url(self.id, getattr(self, "version", "")),
+        }
+
+        job_result = client.job.execute_job(job_kwargs)
+        self.logger.info("job_result: {result}, job_kwargs: {kwargs}".format(result=job_result, kwargs=job_kwargs))
+        if job_result["result"]:
+            job_instance_id = job_result["data"]["job_instance_id"]
+            data.outputs.job_inst_url = get_job_instance_url(biz_cc_id, job_instance_id)
+            data.outputs.job_inst_id = job_instance_id
+            data.outputs.job_inst_name = job_result["data"]["job_instance_name"]
+            data.outputs.client = client
+            return True
+        else:
+            message = job_handle_api_error("job.execute_job", job_kwargs, job_result)
+            self.logger.error(message)
+            data.outputs.ex_data = message
+            return False
 
     def schedule(self, data, parent_data, callback_data=None):
         return super(JobExecuteTaskServiceBase, self).schedule(data, parent_data, callback_data)
